make_hyperstack catch no windwos
For single scans, the section that closes anything with a NFrame == 1 raises an error during the next IJ.getImage() and causes a pop-up. I added a checked for open windows and if there are none, it prints to the log file and moves on before trying to get hte Image.