make_hyperstack, if no windows open, throw exception
Prior to this commit, if there was a single slice movie that was only one frame, it would be closed and then the next section that tried to run IJ.getImage() would fail to find windows and throw a dialogue box with an error. This commit tries to call the WindowManager to get any open window IDs, and if it fails it will raise an exception prior to running imp= IJ.getImage() so no dialogue boxes open.